benny/requirements.txt
alembic==1.0.7
astroid==2.1.0
bcrypt==3.1.6
Cerberus==1.2
cffi==1.12.1
colorama==0.4.1
falcon==1.4.1
falcon-auth==1.1.0
filelock==3.0.10
grpcio==1.19.0
grpcio-tools==1.19.0
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
pluggy==0.8.1
protobuf==3.7.0
py==1.7.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.2.2
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
python-mimeparse==1.6.0
six==1.12.0
SQLAlchemy==1.2.18
toml==0.10.0
tox==3.7.0
virtualenv==16.4.0
wrapt==1.11.1
